trail_2/requirements.txt
flet==0.22.*
requests==2.31.0
pillow==10.2.0